Fixed powerups
Powerups now move mostly as intended (still seems to be a little strange
with the sin curve behavior, will require future fix) and upgrades the
player as expected. Also aligns the new guns with the player's cursor as
expected. Also, the powerup provides a preview of the powerup on itself.
Added 6 different gun presets.