Implement grid controller for Snake
Starting a new mini-project to implement Snake in Unity.
This commit includes a small demo for generating a grid of sprites
which fits to the player's screen. It will adjust the camera to fit
in 40 cells vertically and, using the aspect ratio, will fit in what
it can in the x direction. The x direction will likely not be a
perfect fit, so we position some black bars on the sides as
required. This commit includes some details that are not yet in use
such as CellValidity.