Fix bug where player can turn 180 degrees
It was possible to the player to turn 180 degrees if they pressed
e.g. up, left within a single movement had passed. In this example,
if they were going right, they would go left on the next tick.
Fix the problem by keeping track of the direction that the previous
step was, which we cannot turn 180 degrees from on the next step.