Reset game when player eats own tail
This implements the second failure-condition (in addition to
hitting the wall) - the player eats their own tail.

This implementation is a bit sad as it involves iterating over the
snake body's LinkedList, rather than a smarter O(1) look-up-type
implementation. However, given the limited size of the player's
body (i.e. dozens, 100s of cells at most), the cost is not a
concern. If it becomes one, a better implementation can be done.